Editor and snippet line display instances of CodeMirror have their own respective heights
