+++ src/components/AnnotationCreator.jsx
@@ -1,16 +1,9 @@
 import React, { PropTypes } from 'react';
-import CodeMirror from 'react-codemirror';
 
 import RaisedButton from 'material-ui/RaisedButton';
 import TextField from 'material-ui/TextField';
 
-// Base options for CodeMirror instances for an AnnotationDisplay
-const baseOptions = {
-  lineNumbers: true,
-  theme: 'codesplain',
-  readOnly: true,
-  cursorBlinkRate: -1,
-};
+import LineSnippet from './LineSnippet';
 
 class AnnotationCreator extends React.Component {
   constructor(props) {
@@ -41,18 +34,12 @@ class AnnotationCreator extends React.Component {
   }
 
   render() {
-    const codeMirrorOptions = {
-      ...baseOptions,
-      firstLineNumber: this.props.lineNumber + 1,
-    }
+
     return (
       <div>
-        <CodeMirror
-          style={{
-            height: 'auto',
-          }}
+        <LineSnippet
+          lineNumber={this.props.lineNumber + 1}
           value={this.props.lineText}
-          options={codeMirrorOptions}
         />
         <TextField
           name="annotationEditor"
