+++ src/styles/codesplain.css
@@ -4,7 +4,3 @@
   padding-top: 2px;
   padding-bottom: 2px;
 }
-
-.CodeMirror {
-  height: auto;
-}
